+++ main.go
@@ -1,10 +1,9 @@
 package main
 
 import (
-	"database/sql"
 	"fmt"
 	"log"
-	"tasktracker-api/pkg/models"
+	database "tasktracker-api/db"
 	"tasktracker-api/pkg/repository"
 	router "tasktracker-api/pkg/router"
 	"tasktracker-api/pkg/service"
@@ -19,7 +18,7 @@ func main() {
 	if err != nil {
 		log.Fatalf("error initializing configs: %s", err.Error())
 	}
-	DBConfig := models.DBConfig{
+	DBConfig := database.DBConfig{
 		Username: viper.GetString("db.username"),
 		Password: viper.GetString("db.password"),
 		Host:     viper.GetString("db.host"),
@@ -28,29 +27,15 @@ func main() {
 		SSLmode:  viper.GetString("db.sslmode"),
 	}
 	// Opening a driver typically will not attempt to connect to the database.
-	db, err := sql.Open("postgres", fmt.Sprintf("host=%s port=%s user=%s password=%s dbname=%s sslmode=%s", DBConfig.Host, DBConfig.Port, DBConfig.Username, DBConfig.Password, DBConfig.DBname, DBConfig.SSLmode))
+	db, err := database.NewDatabase(DBConfig)
 	if err != nil {
-		// This will not be a connection error, but a DSN parse error or
-		// another initialization error.
-		log.Fatal(err)
+		log.Fatal(fmt.Printf("init database error : %v", err))
 	}
 	err = db.Ping()
 	if err != nil {
-		log.Fatal(err)
+		log.Fatal(fmt.Printf("ping database error : %v", err))
 	}
 	defer db.Close()
-	//MIGRATIONS
-	// m, err := migrate.New(
-	// 	"file://db/migrations/",
-	// 	"postgres://postgres:secret@localhost:5432/postgres?sslmode=disable")
-	// if err != nil {
-	// 	fmt.Println("migr err")
-	// 	log.Fatal("migrate error: ", err)
-	// }
-	// if err := m.Up(); err != nil {
-	// 	fmt.Println("migr err2")
-	// 	log.Fatal(err)
-	// }
 
 	repository := repository.NewRepository(db)
 	services := service.NewService(repository)
