--- pkg/models/database.go
@@ -1,10 +1,0 @@
-package models
-
-type DBConfig struct {
-	Username string
-	Password string
-	Host     string
-	Port     string
-	DBname   string
-	SSLmode  string
-}
